requirements.txt
cloudpickle==3.1.0
contourpy==1.3.0
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.9.0
greenlet==3.0.3
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.1
matplotlib==3.9.2
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.4.1
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==10.4.0
playwright==1.47.0
PyAutoGUI==0.9.54
pyee==12.0.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.2
setuptools==75.1.0
six==1.16.0
stable_baselines3==2.3.2
sympy==1.13.3
torch==2.4.1
typing_extensions==4.12.2
tzdata==2024.2
